Move contents of api/explorer to api explorer.api namespace package
