+++ api/pyproject.toml
@@ -13,6 +13,7 @@ packages = [{include = "explorer"}]
 [tool.poetry.dependencies]
 python = "^3.10"
 
+click = "^8.1.3"
 gremlinpython = "^3.6" 
 gunicorn = "^20.1.0"
 networkx = "^2.8.7"
@@ -21,6 +22,8 @@ requests = "^2.28.1"
 pydantic = "^1.10.7"
 fastapi = "^0.95.0"
 uvicorn = {extras = ["standard"], version = "^0.21.1"}
+
+nixpkgs-graph-core = { version = "^0.1.0" }
 
 [tool.isort]
 profile = "black"
@@ -31,6 +34,8 @@ black = "^23.1.0"
 flake8 = "^6.0.0"
 pyright = "^1.1.298"
 isort = "^5.12.0"
+
+nixpkgs-graph-core = { path="../core", develop=true }
 
 [build-system]
 requires = ["poetry-core"]
